+++ synapse-edison/src/main/resources/static/js/init-tooltip.js
@@ -1,0 +1,4 @@
+$(document).ready(function () {
+    $('a[data-toggle=tooltip]').tooltip();
+    $('button[data-toggle=tooltip]').tooltip();
+});